+++ src/dominios/Email.cpp
@@ -29,7 +29,7 @@ bool Email::validar(string valor)
         return false;
     }
 
-    for (int i = 0; i < local.length(); i++)
+    for (size_t i = 0; i < local.length(); i++)
     {
         char caractere = local[i];
         if (!StringUtils::isChar(caractere, false) && !StringUtils::isDigit(caractere) && !StringUtils::isIn(caractere, ".-"))
@@ -44,7 +44,7 @@ bool Email::validar(string valor)
     }
 
     string dominio = partes[1];
-    int count = 0;
+    size_t count = 0;
 
     for (char caractere : dominio)
     {
@@ -67,7 +67,7 @@ bool Email::validar(string valor)
             return false;
         }
 
-        for (int i = 0; i < subdominio.length(); i++)
+        for (size_t i = 0; i < subdominio.length(); i++)
         {
             char caractere = subdominio[i];
             if (!StringUtils::isChar(caractere, false) && !StringUtils::isDigit(caractere) && !StringUtils::isIn(caractere, "-"))
